Created the framework to make grpc calls in controller service
