+++ controller/service/pom.xml
@@ -24,19 +24,38 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>mft-api</artifactId>
+        <artifactId>mft-controller</artifactId>
         <groupId>org.apache.airavata</groupId>
         <version>0.01-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>mft-api-service</artifactId>
-
+    <artifactId>mft-controller-service</artifactId>
     <dependencies>
         <dependency>
             <groupId>org.apache.airavata</groupId>
-            <artifactId>mft-api-stub</artifactId>
+            <artifactId>mft-controller-stub</artifactId>
             <version>0.01-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.airavata</groupId>
+            <artifactId>mft-admin</artifactId>
+            <version>0.01-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.orbitz.consul</groupId>
+            <artifactId>consul-client</artifactId>
+            <version>${consul.client}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>log4j-over-slf4j</artifactId>
+            <version>${log4j.over.slf4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
+            <version>${spring.boot.data.jpa}</version>
         </dependency>
         <dependency>
             <groupId>com.h2database</groupId>
@@ -45,34 +64,19 @@
             <scope>runtime</scope>
         </dependency>
         <dependency>
-            <groupId>org.mariadb.jdbc</groupId>
-            <artifactId>mariadb-java-client</artifactId>
-            <version>${mariadb.jdbc}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.airavata</groupId>
-            <artifactId>mft-admin</artifactId>
-            <version>0.01-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.airavata</groupId>
-            <artifactId>mft-core</artifactId>
-            <version>0.01-SNAPSHOT</version>
+            <groupId>net.sf.dozer</groupId>
+            <artifactId>dozer</artifactId>
+            <version>${dozer}</version>
         </dependency>
         <dependency>
             <groupId>javax.xml.bind</groupId>
             <artifactId>jaxb-api</artifactId>
             <version>2.3.1</version>
         </dependency>
-
-        <!-- To be removed -->
-
         <dependency>
-            <groupId>org.apache.airavata</groupId>
-            <artifactId>mft-scp-transport</artifactId>
-            <version>0.01-SNAPSHOT</version>
-            <scope>runtime</scope>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-jpa</artifactId>
+            <version>${spring.boot.data.jpa}</version>
         </dependency>
     </dependencies>
 
@@ -84,17 +88,16 @@
                 <version>${maven.assembly.plugin}</version>
                 <executions>
                     <execution>
-                        <!--FIXME: Was the earlier id an error?-->
-                        <id>mft-api-distribution-package</id>
+                        <id>mft-controller-distribution-package</id>
                         <phase>package</phase>
                         <goals>
                             <goal>single</goal>
                         </goals>
                         <configuration>
                             <tarLongFileMode>posix</tarLongFileMode>
-                            <finalName>${api.service.dist.name}</finalName>
+                            <finalName>${controller.service.dist.name}</finalName>
                             <descriptors>
-                                <descriptor>src/main/assembly/api-service-bin-assembly.xml</descriptor>
+                                <descriptor>src/main/assembly/controller-service-bin-assembly.xml</descriptor>
                             </descriptors>
                             <attach>false</attach>
                         </configuration>
@@ -105,7 +108,8 @@
     </build>
 
     <properties>
-        <api.service.dist.name>API-Service-0.01</api.service.dist.name>
+        <controller.service.dist.name>Controller-Service-0.01</controller.service.dist.name>
         <maven.assembly.plugin>3.1.1</maven.assembly.plugin>
     </properties>
+
 </project>
