+++ controller/service/src/main/java/org/apache/airavata/mft/controller/sql/repository/TransferRequestRepository.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.airavata.mft.controller.sql.repository;
+
+import org.apache.airavata.mft.controller.sql.entity.TransferRequestEntity;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface TransferRequestRepository extends CrudRepository<TransferRequestEntity, String> {
+    Optional<TransferRequestEntity> findByTransferId(String transferId);
+
+    @Modifying
+    @Query("update TransferRequestEntity requestEntity set requestEntity.status = ?2 where requestEntity.transferId = ?1")
+    void updateTransferRequestStatus(String transferId, String status);
+
+    List<TransferRequestEntity> findAll();
+}
