+++ controller/service/src/main/resources/distribution/bin/setenv.sh
