--- controller/src/main/resources/application.properties
@@ -1,26 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-consul.host=localhost
-consul.port=8500
-
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2
-
-spring.datasource.url=jdbc:h2:mem:mft_db;DB_CLOSE_DELAY=-1
-spring.datasource.driveClassName=org.h2.Driver
-spring.datasource.username=airavata
-spring.datasource.password=123456
